SOme progress
